requirements.txt
absl-py==0.11.0
APScheduler==3.6.3
astor==0.8.1
bleach==3.2.1
cached-property==1.5.2
certifi==2020.11.8
cffi==1.14.3
cryptography==3.2.1
decorator==4.4.2
fawkes==0.3.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
importlib-metadata==2.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.3
numpy==1.16.4
opt-einsum==3.3.0
packaging==20.4
Pillow==8.0.1
pip==20.2.4
protobuf==3.13.0
pycparser==2.20
pyparsing==2.4.7
python-telegram-bot==13.0
pytz==2020.4
PyYAML==5.3.1
scipy==1.5.4
setuptools==50.3.2
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tornado==6.1
tzlocal==2.1
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.35.1
wrapt==1.12.1
zipp==3.4.0